remove whole recovery method sheet (recover_sheet.dart)
