+++ lib/utils/guardian_helpers.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:typed_data';
 
 import 'package:biometric_storage/biometric_storage.dart';
@@ -6,24 +7,31 @@ import 'package:candide_mobile_app/models/batch.dart';
 import 'package:candide_mobile_app/models/gnosis_transaction.dart';
 import 'package:candide_mobile_app/models/guardian_operation.dart';
 import 'package:candide_mobile_app/controller/address_persistent_data.dart';
+import 'package:candide_mobile_app/models/recovery_request.dart';
+import 'package:candide_mobile_app/screens/onboard/recovery/recovery_request_page.dart';
 import 'package:candide_mobile_app/services/bundler.dart';
 import 'package:candide_mobile_app/controller/guardian_controller.dart';
 import 'package:candide_mobile_app/controller/settings_persistent_data.dart';
 import 'package:candide_mobile_app/models/relay_response.dart';
 import 'package:candide_mobile_app/screens/home/components/prompt_password.dart';
 import 'package:candide_mobile_app/screens/home/guardians/guardian_operation_review.dart';
+import 'package:candide_mobile_app/services/security.dart';
+import 'package:candide_mobile_app/utils/constants.dart';
 import 'package:candide_mobile_app/utils/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:hive/hive.dart';
 import 'package:magic_sdk/magic_sdk.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
+import 'package:wallet_dart/contracts/factories/EIP4337Manager.g.dart';
 import 'package:wallet_dart/contracts/wallet.dart';
 import 'package:wallet_dart/wallet/UserOperation.dart';
 import 'package:wallet_dart/wallet/wallet_helpers.dart';
+import 'package:wallet_dart/wallet/wallet_instance.dart';
+import 'package:web3dart/crypto.dart';
 import 'package:web3dart/web3dart.dart';
 
-class GuardiansHelper {
+class GuardianOperationsHelper {
 
   static Batch? grantBatch;
 
@@ -118,94 +126,6 @@ class GuardiansHelper {
     Get.back(result: true);
   }
 
-  static void confirmGuardianRecoveryTransaction(List<UserOperation> userOperations) async {
-    List<UserOperation>? unsignedUserOperations = await Bundler.requestPaymasterSignature(
-      userOperations,
-      SettingsData.network,
-    );
-    if (unsignedUserOperations == null) return;
-    if (!Bundler.verifyUserOperationsWithPaymaster(userOperations, unsignedUserOperations)) {
-      Get.back();
-      Utils.showError(title: "Error", message: "Transaction corrupted, contact us for help");
-      return;
-    }
-    var signedUserOperations = await Bundler.signUserOperationsAsMagicLink(
-      SettingsData.network,
-      unsignedUserOperations,
-    );
-
-    if (signedUserOperations == null){
-      Utils.showError(title: "Error", message: "Incorrect password");
-      return;
-    }
-
-    BotToast.showText(
-      text: "Transaction sent, this might take a minute...",
-      textStyle: const TextStyle(fontWeight: FontWeight.bold, color: Colors.black),
-      contentColor: Get.theme.colorScheme.primary,
-      align: Alignment.topCenter,
-    );
-    var cancelLoad = Utils.showLoading();
-    //
-    RelayResponse? response = await Bundler.relayUserOperations(signedUserOperations, SettingsData.network);
-    if (response?.status == "PENDING"){
-      BotToast.showText(
-        text: "Transaction still pending, refresh later...",
-        textStyle: const TextStyle(fontWeight: FontWeight.bold, color: Colors.black),
-        contentColor: Get.theme.colorScheme.primary,
-        align: Alignment.topCenter,
-      );
-    }else if (response?.status == "FAIL"){
-      BotToast.showText(
-        text: "Transaction failed, contact us for help",
-        textStyle: const TextStyle(fontWeight: FontWeight.bold),
-        contentColor: Colors.red,
-        align: Alignment.topCenter,
-      );
-    }else if (response?.status == "SUCCESS"){
-      BotToast.showText(
-        text: "Transaction completed!",
-        textStyle: const TextStyle(fontWeight: FontWeight.bold, color: Colors.white),
-        contentColor: Colors.green,
-        align: Alignment.topCenter,
-      );
-    }
-    cancelLoad();
-    Get.back(result: true);
-  }
-
-
-  /*static Future<bool> revokeGuardian(String address) async {
-    CancelFunc? cancelLoad = Utils.showLoading();
-    Map data = await GuardianController.buildRevokeOps(
-      instance: AddressData.wallet,
-      network: SettingsData.network,
-      isDeployed: AddressData.walletStatus.proxyDeployed,
-      nonce: AddressData.walletStatus.nonce,
-      defaultCurrency: SettingsData.quoteCurrency,
-      guardianAddress: address,
-    );
-    cancelLoad();
-    cancelLoad = null;
-    bool? refresh = await showBarModalBottomSheet(
-      context: Get.context!,
-      builder: (context) {
-        Get.put<ScrollController>(ModalScrollController.of(context)!, tag: "guardian_review_modal");
-        return GuardianOperationReview(
-          operation: GuardianOperation.revoke,
-          guardian: address,
-          fee: data["fee"],
-          onConfirm: (){
-            cancelLoad = Utils.showLoading();
-            onConfirmTransaction(data["userOperations"]);
-          },
-        );
-      },
-    );
-    cancelLoad?.call();
-    return (refresh ?? false);
-  }*/
-
   static Future<bool> grantGuardian(String address, {Map? magicLinkData}) async {
     CancelFunc? cancelLoad = Utils.showLoading();
     Batch grantBatch = Batch();
@@ -213,20 +133,13 @@ class GuardiansHelper {
     int friendsCount = 0;
     int threshold = 1;
     if (AddressData.walletStatus.socialModuleDeployed){
-      print("G!");
       EthereumAddress recoveryManager = await CWallet.recoveryInterface(AddressData.wallet.socialRecovery).manager();
-      print("G! : ${recoveryManager.hex}");
-      print("G! : ${AddressData.wallet.walletAddress.hex}");
       if (recoveryManager.hex == AddressData.wallet.walletAddress.hex){
-        print("G!!");
         setupSocialModule = false;
         friendsCount = (await CWallet.recoveryInterface(AddressData.wallet.socialRecovery).friendsCount()).toInt();
         threshold = (((friendsCount + 1) / 2 ) + 1).floor();
       }
     }
-    print("Setup: $setupSocialModule");
-    print("friendsCount: $friendsCount");
-    print("threshold: $threshold");
     List<GnosisTransaction> transactions = GuardianController.buildGrantTransactions(
         socialModuleDeployed: AddressData.walletStatus.socialModuleDeployed,
         setup: setupSocialModule,
@@ -316,60 +229,6 @@ class GuardiansHelper {
     return (refresh ?? false);
   }
 
-  /*static Future<bool> recoverUsingGuardian(String address, String newOwner) async {
-    CancelFunc? cancelLoad = Utils.showLoading();
-    Map data = await GuardianController.buildRecoverOps(
-      walletAddress: EthereumAddress.fromHex(address),
-      network: SettingsData.network,
-      defaultCurrency: SettingsData.quoteCurrency,
-      newOwner: newOwner,
-    );
-    cancelLoad();
-    cancelLoad = null;
-    bool? refresh = await showBarModalBottomSheet(
-      context: Get.context!,
-      builder: (context) {
-        Get.put<ScrollController>(ModalScrollController.of(context)!, tag: "guardian_review_modal");
-        return GuardianOperationReview(
-          operation: GuardianOperation.recover,
-          guardian: address,
-          fee: data["fee"],
-          onConfirm: (){
-            cancelLoad = Utils.showLoading();
-            confirmGuardianRecoveryTransaction(data["userOperations"]);
-          },
-        );
-      },
-    );
-    cancelLoad?.call();
-    return (refresh ?? false);
-  }*/
-
-  /*static Future<bool> setupMagicLinkRecovery(String email, String walletAddress, String newOwner) async {
-    CancelFunc? cancelLoad = Utils.showLoading();
-    bool? refresh = false;
-    try {
-      var magic = Magic.instance;
-      var isLoggedIn = await magic.user.isLoggedIn();
-      if (isLoggedIn){
-        await magic.user.logout();
-      }
-      await magic.auth.loginWithMagicLink(email: email, showUI: true); // todo show ui
-      var metadata = await magic.user.getMetadata();
-      cancelLoad.call();
-      cancelLoad = null;
-      if (metadata.publicAddress == null) return false;
-      refresh = await recoverUsingGuardian(walletAddress, newOwner);
-    } on Exception catch (e) {
-      cancelLoad?.call();
-      print(e);
-      return false;
-    }
-    cancelLoad?.call();
-    Get.back();
-    return refresh;
-  }*/
-
   static Future<bool> setupMagicLinkGuardian(String email) async {
     CancelFunc? cancelLoad = Utils.showLoading();
     bool? refresh = false;
@@ -394,5 +253,120 @@ class GuardiansHelper {
     Get.back();
     return refresh;
   }
+}
 
+class GuardianRecoveryHelper{
+
+  static Future<EthereumAddress?> getSocialRecoveryModule(String walletAddress, [GetModulesPaginated? modulesPaginated]) async{
+    if (modulesPaginated == null){
+      try{
+        modulesPaginated = await CWallet.customInterface(EthereumAddress.fromHex(walletAddress)).getModulesPaginated(Constants.addressOne, BigInt.from(25));
+      }catch (e){
+        return null;
+      }
+    }
+    for (EthereumAddress module in modulesPaginated.array){
+      try{
+        int guardiansCount = (await CWallet.recoveryInterface(module).friendsCount()).toInt();
+        if (guardiansCount == 0) return null;
+        return module;
+      }catch (e){
+        continue;
+      }
+    }
+    return null;
+  }
+
+  static Future<EthereumAddress?> getModuleManagerAddress(String walletAddress, [GetModulesPaginated? modulesPaginated]) async{
+    if (modulesPaginated == null){
+      try{
+        modulesPaginated = await CWallet.customInterface(EthereumAddress.fromHex(walletAddress)).getModulesPaginated(Constants.addressOne, BigInt.from(25));
+      }catch (e){
+        return null;
+      }
+    }
+    String abi = '''[{"inputs":[],"name":"eip4337manager","outputs":[{"internalType":"address","name":"","type":"address"}],"stateMutability":"view","type":"function"}]''';
+    for (EthereumAddress module in modulesPaginated.array){
+      try{
+        final contract = DeployedContract(ContractAbi.fromJson(abi, 'EIP4337Fallback'), module);
+        var moduleManagerAddress = await Constants.client.call(contract: contract, function: contract.function("eip4337manager"), params: []);
+        print("Fallback: ${module.hexEip55}");
+        return moduleManagerAddress[0];
+      }catch (e){
+        continue;
+      }
+    }
+    return null;
+  }
+
+  static setupRecoveryWallet(String address, String password, bool biometricsEnabled, String method) async {
+    var cancelLoad = Utils.showLoading();
+    //
+    GetModulesPaginated modulesPaginated;
+    try{
+      modulesPaginated = await CWallet.customInterface(EthereumAddress.fromHex(address)).getModulesPaginated(Constants.addressOne, BigInt.from(25));
+    }catch (e){
+      cancelLoad();
+      Utils.showError(title: "Error", message: "This wallet is not a candide smart contract wallet!");
+      return;
+    }
+    //
+    EthereumAddress? socialModuleAddress = await getSocialRecoveryModule(address, modulesPaginated);
+    EthereumAddress? moduleManagerAddress = await getModuleManagerAddress(address, modulesPaginated);
+    print(socialModuleAddress?.hexEip55);
+    print(moduleManagerAddress?.hexEip55);
+    if (socialModuleAddress == null || moduleManagerAddress == null){
+      cancelLoad();
+      Utils.showError(title: "Error", message: "This wallet does not have any guardians, unfortunately this means this wallet cannot be recovered, contact us to learn more");
+      return;
+    }
+    //
+    if (biometricsEnabled){
+      try {
+        final store = await BiometricStorage().getStorage('auth_data');
+        await store.write(password);
+        await Hive.box("settings").put("biometrics_enabled", true);
+      } on AuthException catch(_) {
+        BotToast.showText(
+            text: "User cancelled biometrics auth, please try again",
+            contentColor: Colors.red.shade900,
+            align: Alignment.topCenter,
+            borderRadius: BorderRadius.circular(20)
+        );
+        return;
+      }
+    }else{
+      await Hive.box("settings").put("biometrics_enabled", false);
+    }
+    var salt = base64Encode(Utils.randomBytes(16, secure: true));
+    WalletInstance wallet = await WalletHelpers.createRecovery(address, moduleManagerAddress.hexEip55, socialModuleAddress.hexEip55, password, salt);
+    await Hive.box("wallet").put("recovered", jsonEncode(wallet.toJson()));
+    cancelLoad();
+    Get.back();
+    //
+    if (method == "social-recovery"){
+      EthereumAddress oldOwner = (await CWallet.customInterface(EthereumAddress.fromHex(address)).getOwners())[0];
+      var dataHash = CWallet.customInterface(EthereumAddress.fromHex(address)).self.function("swapOwner").encodeCall([Constants.addressOne, oldOwner, EthereumAddress.fromHex(wallet.initOwner)]);
+      dataHash = keccak256(dataHash);
+      startRecoveryRequest(wallet, bytesToHex(dataHash, include0x: true), oldOwner.hexEip55);
+    }
+    //
+  }
+
+  static void startRecoveryRequest(WalletInstance wallet, String dataHash, String oldOwner) async {
+    var cancelLoad = Utils.showLoading();
+    RecoveryRequest? request = await SecurityGateway.create(wallet.walletAddress.hexEip55, wallet.socialRecovery.hexEip55, dataHash, oldOwner, wallet.initOwner, SettingsData.network);
+    if (request == null){
+      cancelLoad();
+      if (SecurityGateway.latestErrorCode == 429){
+        Utils.showError(title: "Error", message: "A recovery request was created for this wallet in the past hour, please wait some time and try again");
+      }else{
+        Utils.showError(title: "Error", message: "Error occurred while trying to create a recovery request, please try again later or contact us");
+      }
+      return;
+    }
+    cancelLoad();
+    AddressData.storeRecoveryRequest(request.id);
+    Get.off(RecoveryRequestPage(request: request));
+  }
 }
